requirements.txt
# =============================================================================
# Al-Muhami Al-Zaki — MVP Dependencies
# Python 3.10+
# =============================================================================

# -----------------------------------------------------------------------------
# Core Framework
# -----------------------------------------------------------------------------
langchain>=0.3.0
langchain-core>=0.3.0
langgraph>=0.2.0
pydantic>=2.0.0

# -----------------------------------------------------------------------------
# Vector Database (Qdrant)
# -----------------------------------------------------------------------------
qdrant-client>=1.12.0

# -----------------------------------------------------------------------------
# Embeddings
# -----------------------------------------------------------------------------
sentence-transformers>=3.0.0
transformers>=4.40.0
torch>=2.0.0

# -----------------------------------------------------------------------------
# LLM Providers (Free Tier)
# -----------------------------------------------------------------------------
langchain-groq>=0.2.0           # Llama-3 for grading (API backup)
langchain-google-genai>=2.0.0   # Gemini Flash (API backup)
langchain-ollama>=0.2.0         # Ollama for local LLM (unlimited)

# -----------------------------------------------------------------------------
# Document Processing
# -----------------------------------------------------------------------------
pypdf>=4.0.0                    # PDF text extraction
pdfplumber>=0.11.0              # Advanced PDF parsing (tables)
python-docx>=1.1.0              # Word document support

# -----------------------------------------------------------------------------
# Arabic NLP (Law 151 Compliance)
# -----------------------------------------------------------------------------
camel-tools>=1.5.0              # CAMeL Arabic NLP toolkit
pyarabic>=0.6.15                # Arabic text utilities

# -----------------------------------------------------------------------------
# UI
# -----------------------------------------------------------------------------
streamlit>=1.40.0

# -----------------------------------------------------------------------------
# Utilities
# -----------------------------------------------------------------------------
python-dotenv>=1.0.0            # Environment management
loguru>=0.7.0                   # Structured logging
tenacity>=8.0.0                 # Retry logic for API calls
httpx>=0.27.0                   # Async HTTP client

# -----------------------------------------------------------------------------
# Testing & Quality
# -----------------------------------------------------------------------------
pytest>=8.0.0
pytest-asyncio>=0.24.0
ragas>=0.2.0                    # RAG evaluation

# -----------------------------------------------------------------------------
# Development
# -----------------------------------------------------------------------------
black>=24.0.0
ruff>=0.5.0
mypy>=1.10.0
